upgrade spring and project version to 3.0.2, fix all use of spring and spring batch on the project, removed "map repository" plugin (removed from spring batch)

+++ spring-batch-lightmin-api/src/main/java/org/tuxdevelop/spring/batch/lightmin/api/resource/monitoring/StepExecutionEventInfo.java
@@ -12,11 +12,11 @@ public class StepExecutionEventInfo {
     private String stepName;
     private ExitStatus exitStatus;
 
-    private int readCount;
-    private int writeCount;
-    private int commitCount;
-    private int rollbackCount;
-    private int readSkipCount;
-    private int processSkipCount;
-    private int writeSkipCount;
+    private long readCount;
+    private long writeCount;
+    private long commitCount;
+    private long rollbackCount;
+    private long readSkipCount;
+    private long processSkipCount;
+    private long writeSkipCount;
 }
